purchase working but not insert well
