+++ inventory-management/src/hooks/useMediaQuery.ts
@@ -1,0 +1,28 @@
+// src/hooks/useMediaQuery.ts
+import { useState, useEffect } from 'react';
+
+const useMediaQuery = (query: string): boolean => {
+  const [matches, setMatches] = useState(false);
+
+  useEffect(() => {
+    const media = window.matchMedia(query);
+    
+    // Update state with current match
+    if (media.matches !== matches) {
+      setMatches(media.matches);
+    }
+    
+    // Create listener that updates state on change
+    const listener = () => setMatches(media.matches);
+    
+    // Add listener
+    media.addEventListener('change', listener);
+    
+    // Clean up
+    return () => media.removeEventListener('change', listener);
+  }, [matches, query]);
+
+  return matches;
+};
+
+export default useMediaQuery;
